Created infrastructure for java21
* Run CDK in java 21 with image `ghcr.io/muhamadto/spring-native-amazonlinux2-builder:21-amazonlinux2`
* Replaced `com.coffeebeans.springnativeawslambda.infra.lambda.CustomRuntime2Function` with `com.coffeebeans.springnativeawslambda.infra.lambda.CustomRuntime2023FunctionJava21` and updated tests.
* `docker-compose.yml` now adopting `apigateway` functions from `ghcr.io/muhamadto/spring-native-amazonlinux2-builder:21-amazonlinux2-awscliv2`. However, the lambda is not loading and need some more work.
* Reinstated `native:compile` in `prepare-package` phase. This is to ensure the `spring-native-aws-lambda-function-native-zip.zip` include would include the `spring-native-aws-lambda-function` native executable.
