+++ .dockerignore
@@ -2,6 +2,8 @@
 .idea/
 badges/
 **/target/
+cdk.out/
+volume/
 .gitignore
 CODE_OF_CONDUCT.md
 HELP.md
