+++ spring-native-aws-lambda-function/pom.xml
@@ -167,7 +167,7 @@
                 <buildArg>--verbose</buildArg>
                 <buildArg>--no-fallback</buildArg>
                 <buildArg>--enable-preview</buildArg>
-                <buildArg>--gc=G1</buildArg>
+                  <buildArg>--gc=G1</buildArg>
                 <buildArg>-march=native</buildArg>
                 <buildArg>--strict-image-heap</buildArg>
                 <buildArg>--enable-url-protocols=http</buildArg>
@@ -175,6 +175,15 @@
                 <buildArg>-H:+ReportExceptionStackTraces</buildArg>
               </buildArgs>
             </configuration>
+            <executions>
+              <execution>
+                <id>build-native</id>
+                <phase>prepare-package</phase>
+                <goals>
+                  <goal>compile</goal>
+                </goals>
+              </execution>
+            </executions>
           </plugin>
           <plugin>
             <artifactId>maven-assembly-plugin</artifactId>
