+++ spring-native-aws-lambda-infra/src/main/java/com/coffeebeans/springnativeawslambda/infra/CoffeeBeansConstruct.java
@@ -1,0 +1,8 @@
+package com.coffeebeans.springnativeawslambda.infra;
+
+import software.constructs.IConstruct;
+
+public interface CoffeeBeansConstruct extends IConstruct {
+
+
+}
