+++ spring-native-aws-lambda-infra/src/main/java/com/coffeebeans/springnativeawslambda/infra/Constants.java
@@ -1,0 +1,14 @@
+package com.coffeebeans.springnativeawslambda.infra;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class Constants {
+    public static final String KEY_ENV = "ENV";
+    public static final String KEY_COST_CENTRE = "COST_CENTRE";
+    public static final String KEY_APPLICATION_NAME = "applicationName";
+
+    public static final String VALUE_COST_CENTRE = "coffeeBeans-core";
+
+}
